% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irtoys-package.R
\docType{data}
\name{b2}
\alias{b2}
\title{Example item parameters 2PL (Bilog)}
\format{
The object returned by \code{est}
}
\description{
Item parameter estimates for the 2PL model, estimated with \code{bilog} from
the example data set \code{Scored}. These are provided because not users will
have BILOG-MG available.
}
\keyword{datasets}
