% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracelines.R
\name{erf}
\alias{erf}
\title{Plot empirical response function}
\usage{
erf(resp, item)
}
\arguments{
\item{resp}{A matrix of binary responses to a test, with persons as rows and
items as columns.}

\item{item}{The item (column number in \code{resp}) to plot.}
}
\value{
None
}
\description{
Produces a plot of the empirical response function (proportion of correct 
responses at each possible sum score).
}
\examples{

erf(Scored, 3)

}
\author{
Ivailo Partchev
}
\keyword{models}

