% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{interactionModel}
\alias{interactionModel}
\title{Interaction model}
\usage{
interactionModel(x, eps = 0.1, rasch = FALSE)
}
\arguments{
\item{x}{A matrix of responses: persons as rows, items as columns,
entries are either 0 or 1, no missing data}

\item{eps}{Convergence criterion, default is 1e-1}

\item{rasch}{if TRUE, the Rasch model will be estimated instead.}
}
\value{
A matrix
}
\description{
Estimates the parameters of the interaction model. The plots can be produced with the
appropriate plotting method.
}
\references{
Shelby Haberman. The interaction model. Chapter in: 
Mathias von Davier & Claus H. Carstensen (Eds.) (2007). 
Multivariate and Mixture Distribution Rasch Models. New York:Springer
}
\author{
Ivailo Partchev, using C code and theory by Gunter Maris
}
\keyword{models}
