% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracelines.R
\name{scoreMetric}
\alias{scoreMetric}
\title{Sum score metric}
\usage{
scoreMetric(ip)
}
\arguments{
\item{ip}{Item parameters: the output of \code{est}, or a 3-column matrix 
corresponding to its first element, \code{est}.}
}
\value{
An object with the same structure as the output of \code{irf}
}
\description{
From a set of estimated item parameters, compute an item response function in sum score
metric, i.e., the probabilities of a correct response predicted by the model at abilities
corresponding to the possible (raw) sum scores.
}
\author{
Ivailo Partchev
}
\keyword{models}
