% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tet}
\alias{tet}
\title{Approximate tetchoric correlation matrix}
\usage{
tet(d)
}
\arguments{
\item{d}{a matrix (or data frame, which will be converted to a matrix) containing
only zeroes an ones. NAs are not allowed.}
}
\value{
A matrix of approximate tetchoric correlations.
}
\description{
Matrix of tetchoric correlations using the approximation by 
Bonett and Price (2005).
}
\examples{

tets <- tet(Scored)

}
\references{
Douglas G. Bonett and Robert M. Price (2005). Inferential Methods for the 
Tetrachoric Correlation Coefficient. Journal of Educational and Behavioral Statistics,
Vol. 30, No. 2, pp. 213--225
}
\author{
Ivailo Partchev
}
\keyword{models}
