% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irtoys-package.R
\docType{data}
\name{Unscored}
\alias{Unscored}
\title{Original, unscored multiple-choice responses to a test}
\format{A data set with 472 persons and 18 items. Each item has 4 possible
answers, of which only one is true. There are many NA, which can be treated
as wrong responses.}
\description{
Real-life data set containing the responses to a test, before they have been
recoded as true or false. Can be used with only two functions in the
package: \code{scoreResponses} and \code{tgf}. All other functions expect binary data,
which can be produced with \code{scoreResponses}.
}
\keyword{datasets}
