\name{internal}
\alias{internal}
\alias{est.blm}
\alias{est.icl}
\alias{est.ltm}
\alias{eap.one}
\alias{simple.scale}
\alias{adv.scale}
\alias{hb}
\alias{sl}
\alias{llf}
\title{Internal functions}
\description{These functions are for internal use.}
\author{Ivailo Partchev}
\keyword{models}
