\docType{data}
\name{Scored2pl}
\alias{Scored2pl}
\title{Example item parameters}
\format{A list of two matrices: \code{est} contains the parameter estimates,
and \code{se} contains the standard errors (see also \code{est}).}
\description{
  Item parameter estimates for the 2PL model, estimated
  with \code{ltm} from the example data set \code{Scored}.
  These are provided as a check, and to speed up the
  examples for the various functions in the package.
}
\keyword{datasets}

